chore: modal 위치 변경 및 사용자 측에서 각각의 모달 사용
