+++ components/src/lib/complexModal/PromptModal.tsx
@@ -1,9 +1,11 @@
-import React from 'react';
-import { Modal } from '.';
+import { Modal } from '..';
 
-interface AlertModalProps {
+interface PromptModalProps {
   isOpen: boolean;
+  value: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
   onCloseIcon?: () => void;
+  onCloseButton?: () => void;
   onConfirmButton?: () => void;
   onDimmedClick?: () => void;
   position?: 'center' | 'bottom';
@@ -13,12 +15,15 @@ interface AlertModalProps {
   animationDuration?: number;
   title?: string;
   showCloseIcon?: boolean;
-  buttonLabel?: string;
+  closeButtonLabel?: string;
+  confirmButtonLabel?: string;
+  placeholder?: string;
 }
 
-const AlertModal = ({
+const PromptModal = ({
   isOpen,
   onCloseIcon,
+  onCloseButton,
   onConfirmButton,
   onDimmedClick,
   position = 'center',
@@ -28,8 +33,12 @@ const AlertModal = ({
   animationDuration,
   title,
   showCloseIcon = true,
-  buttonLabel = '확인',
-}: AlertModalProps) => {
+  closeButtonLabel = '취소',
+  confirmButtonLabel = '확인',
+  value,
+  onChange,
+  placeholder = '입력해주세요',
+}: PromptModalProps) => {
   const handleClickDimmed = () => {
     onDimmedClick && onDimmedClick();
   };
@@ -40,6 +49,10 @@ const AlertModal = ({
 
   const handleConfirmButton = () => {
     onConfirmButton && onConfirmButton();
+  };
+
+  const handleCloseButton = () => {
+    onCloseButton && onCloseButton();
   };
 
   return (
@@ -55,12 +68,21 @@ const AlertModal = ({
         <Modal.Title title={title} />
         <Modal.CloseIcon onClose={handleCloseIcon} showCloseIcon={showCloseIcon} />
       </Modal.Header>
-      {content && <Modal.Content>{content}</Modal.Content>}
+      {content && (
+        <Modal.Content>
+          <Modal.Input value={value} onChange={onChange} placeholder={placeholder} autoFocus />
+        </Modal.Content>
+      )}
       <Modal.Footer position="row" justifyContent="flex-end">
-        <Modal.ConfirmButton size="small" label={buttonLabel} onConfirm={handleConfirmButton} />
+        <Modal.CloseButton size="small" label={closeButtonLabel} onClose={handleCloseButton} />
+        <Modal.ConfirmButton
+          size="small"
+          label={confirmButtonLabel}
+          onConfirm={handleConfirmButton}
+        />
       </Modal.Footer>
     </Modal>
   );
 };
 
-export default AlertModal;
+export default PromptModal;
