+++ components/src/lib/index.ts
@@ -1,1 +1,4 @@
 export { default as Modal } from './ModalMain';
+export { default as AlertModal } from './complexModal/AlertModal';
+export { default as ConfirmModal } from './complexModal/ConfirmModal';
+export { default as PromptModal } from './complexModal/PromptModal';
